+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/mysql/v200/Migration.java
@@ -12,7 +12,7 @@ import static org.openmetadata.service.migration.utils.v200.MigrationUtil.migrat
 import lombok.SneakyThrows;
 import org.openmetadata.service.migration.api.MigrationProcessImpl;
 import org.openmetadata.service.migration.utils.MigrationFile;
-import org.openmetadata.service.migration.utils.v200.TaskWorkflowMigrationUtil;
+import org.openmetadata.service.migration.utils.v200.MigrationUtil;
 
 public class Migration extends MigrationProcessImpl {
 
@@ -38,9 +38,9 @@ public class Migration extends MigrationProcessImpl {
     addCreateTaskRuleToDataConsumerPolicy(collectionDAO);
 
     initializeWorkflowHandler();
-    TaskWorkflowMigrationUtil taskWorkflowMigrationUtil = new TaskWorkflowMigrationUtil(handle);
-    taskWorkflowMigrationUtil.runTaskWorkflowCutoverMigration();
-    taskWorkflowMigrationUtil.addTaskResourceToMentionAlerts();
-    taskWorkflowMigrationUtil.runRecognizerFeedbackTaskTypeMigration();
+    MigrationUtil.TaskWorkflow taskWorkflow = new MigrationUtil.TaskWorkflow(handle);
+    taskWorkflow.runTaskWorkflowCutoverMigration();
+    taskWorkflow.addTaskResourceToMentionAlerts();
+    taskWorkflow.runRecognizerFeedbackTaskTypeMigration();
   }
 }
